factor ValuesTrait into ValuesInterface; bugfixes
