+++ src/Common/InsertInterface.php
@@ -19,7 +19,7 @@ use Aura\Sql_Query\QueryInterface;
  * @package Aura.Sql_Query
  *
  */
-interface InsertInterface extends QueryInterface
+interface InsertInterface extends QueryInterface, ValuesInterface
 {
     /**
      *
@@ -31,40 +31,4 @@ interface InsertInterface extends QueryInterface
      *
      */
     public function into($into);
-    
-    /**
-     * 
-     * Sets one column value placeholder.
-     * 
-     * @param string $col The column name.
-     * 
-     * @return $this
-     * 
-     */
-    public function col($col);
-
-    /**
-     * 
-     * Sets multiple column value placeholders.
-     * 
-     * @param array $cols A list of column names.
-     * 
-     * @return $this
-     * 
-     */
-    public function cols(array $cols);
-
-    /**
-     * 
-     * Sets a column value directly; the value will not be escaped, although
-     * fully-qualified identifiers in the value will be quoted.
-     * 
-     * @param string $col The column name.
-     * 
-     * @param string $value The column value expression.
-     * 
-     * @return $this
-     * 
-     */
-    public function set($col, $value);
 }
