+++ src/Common/UpdateInterface.php
@@ -20,7 +20,7 @@ use Aura\Sql_Query\Common\WhereInterface;
  * @package Aura.Sql_Query
  *
  */
-interface UpdateInterface extends QueryInterface, WhereInterface
+interface UpdateInterface extends QueryInterface, WhereInterface, ValuesInterface
 {
     /**
      *
@@ -32,40 +32,4 @@ interface UpdateInterface extends QueryInterface, WhereInterface
      *
      */
     public function table($table);
-    
-    /**
-     * 
-     * Sets one column value placeholder.
-     * 
-     * @param string $col The column name.
-     * 
-     * @return $this
-     * 
-     */
-    public function col($col);
-
-    /**
-     * 
-     * Sets multiple column value placeholders.
-     * 
-     * @param array $cols A list of column names.
-     * 
-     * @return $this
-     * 
-     */
-    public function cols(array $cols);
-
-    /**
-     * 
-     * Sets a column value directly; the value will not be escaped, although
-     * fully-qualified identifiers in the value will be quoted.
-     * 
-     * @param string $col The column name.
-     * 
-     * @param string $value The column value expression.
-     * 
-     * @return $this
-     * 
-     */
-    public function set($col, $value);
 }
